+++ addon/mocks/attribute-matcher.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import FactoryGuy from '../factory-guy';
-import {isEmptyObject, isEquivalent} from '../utils/helper-functions';
-//const { isEmpty } = Ember;
+import {isEquivalent} from '../utils/helper-functions';
+const { isEmpty } = Ember;
 
 /**
  This is a mixin used by MockUpdate and MockCreateRequest
@@ -39,14 +39,13 @@ const AttributeMatcher = (superclass) => class extends superclass {
     );
   }
   
-  extraRequestMatches(request) {
+  extraRequestMatches(settings) {
     if (this.matchArgs) {
-      let requestBody = JSON.parse(request.requestBody),
-          requestData = requestBody[this.modelName];
+      let requestData = JSON.parse(settings.data);
       if (typeof this.matchArgs === 'function') {
         return this.matchArgs(requestData);
       } else {
-        return this.attributesMatch(requestBody);
+        return this.attributesMatch(requestData);
       }
     }
     return true;
@@ -71,26 +70,24 @@ const AttributeMatcher = (superclass) => class extends superclass {
    */
   attributesMatch(requestData) {
     let matchArgs = this.matchArgs;
-    if (isEmptyObject(matchArgs)) {
+    if (isEmpty(Object.keys(matchArgs))) {
       return true;
     }
 
     let builder = FactoryGuy.fixtureBuilder(this.modelName);
 
     // transform they match keys
-//    console.log('matchArgs',matchArgs,Object.keys(matchArgs));
     let matchCheckKeys = Object.keys(matchArgs).map((key)=> {
       return builder.transformKey(this.modelName, key);
     });
-//    console.log('matchCheckKeys',matchCheckKeys);
+
     // build the match args into a JSONPayload class
     let buildOpts = { serializeMode: true, transformKeys: true };
     let expectedData = builder.convertForBuild(this.modelName, matchArgs, buildOpts);
 
     // wrap request data in a JSONPayload class
     builder.wrapPayload(this.modelName, requestData);
-//    console.log('requestData',requestData);
-//    console.log('expectedData',requestData.get());
+
     // success if all values match
     return matchCheckKeys.every((key)=> {
       return isEquivalent(expectedData.get(key), requestData.get(key));
