+++ addon/mocks/maybe-id-url-match.js
@@ -19,19 +19,7 @@ const MaybeIdUrlMatch = (superclass) => class extends superclass {
       return this.id;
     }
   }
-
-  /**
-   *
-   * @returns {String} url
-   */
-  getUrl() {
-    let url = super.getUrl();
-    if (!this.id) {
-      url = `${url}/:id`;
-    }
-    return url;
-  }
-
+  
   urlMatch(settings) {
     /**
      If no id is specified, match any url with an id,
