+++ addon/mocks/mock-delete-request.js
@@ -5,7 +5,6 @@ export default class MockDeleteRequest extends MaybeIdUrlMatch(MockRequest) {
   constructor(modelName, id) {
     super(modelName, 'deleteRecord');
     this.id = id;
-    this.setupHandler();
   }
 
   getType() {
