+++ package.json
@@ -25,8 +25,7 @@
     "broccoli-funnel": "^1.0.1",
     "broccoli-merge-trees": "^1.1.1",
     "ember-cli-babel": "^6.7.0",
-    "jquery-mockjax": "^2.2.1",
-    "pretender": "^1.6.0"
+    "jquery-mockjax": "^2.2.1"
   },
   "devDependencies": {
     "active-model-adapter": "2.1.1",
@@ -37,7 +36,6 @@
     "ember-cli-htmlbars": "1.3.0",
     "ember-cli-htmlbars-inline-precompile": "0.4.0",
     "ember-cli-inject-live-reload": "^1.7.0",
-    "ember-cli-pretender": "^1.0.1",
     "ember-cli-qunit": "^4.0.0",
     "ember-cli-shims": "1.2.0-beta.2",
     "ember-cli-sri": "^2.1.1",
